renamer updated to work for all file extensions and folders automatically

+++ renamer.py
@@ -1,5 +1,6 @@
 import os
 import sys
+import re
 
 new_name=""
 if len(sys.argv)==2:
@@ -7,11 +8,19 @@ if len(sys.argv)==2:
 
 while new_name:
     folder=os.getcwd()
-    for count,filename in enumerate(os.listdir(folder)):
+    count=0
+    for filename in os.listdir(folder):
         if "renamer" in filename:
             continue
         src=f"{folder}/{filename}"
-        dest=f"{folder}/{new_name}{str(count)}.jpg"
+        if not os.path.isfile(src):
+            continue
+        count=count+1
+        ext=""
+        result=re.search(r'\.([a-zA-Z0-9]+)$', filename)
+        if result:
+            ext=ext+result.group(0)
+        dest=f"{folder}/{new_name}{str(count)}"+ ext
         os.rename(src,dest)
     break
